+++ sagemaker-spark-sdk/src/main/scala/com/amazonaws/services/sagemaker/sparksdk/algorithms/ImageURIProvider.scala
@@ -38,7 +38,10 @@ private[algorithms] object SagerMakerRegionAccountMaps {
     Regions.US_EAST_2.getName -> "404615174143",
     Regions.US_WEST_2.getName -> "174872318107",
     Regions.AP_NORTHEAST_1.getName -> "351501993468",
-    Regions.AP_NORTHEAST_2.getName -> "835164637446"
+    Regions.AP_NORTHEAST_2.getName -> "835164637446",
+    Regions.EU_CENTRAL_1.getName -> "664544806723",
+    Regions.AP_SOUTHEAST_2.getName -> "712309505854",
+    Regions.GovCloud.getName -> "226302683700"
   )
 
   // For LDA
@@ -48,7 +51,9 @@ private[algorithms] object SagerMakerRegionAccountMaps {
     Regions.US_EAST_2.getName -> "999911452149",
     Regions.US_WEST_2.getName -> "266724342769",
     Regions.AP_NORTHEAST_1.getName -> "258307448986",
-    Regions.AP_NORTHEAST_2.getName -> "293181348795"
+    Regions.AP_NORTHEAST_2.getName -> "293181348795",
+    Regions.EU_CENTRAL_1.getName -> "353608530281",
+    Regions.AP_SOUTHEAST_2.getName -> "297031611018"
   )
 
   // For XGBoost
@@ -58,7 +63,10 @@ private[algorithms] object SagerMakerRegionAccountMaps {
     Regions.US_EAST_2.getName -> "825641698319",
     Regions.US_WEST_2.getName -> "433757028032",
     Regions.AP_NORTHEAST_1.getName -> "501404015308",
-    Regions.AP_NORTHEAST_2.getName -> "306986355934"
+    Regions.AP_NORTHEAST_2.getName -> "306986355934",
+    Regions.EU_CENTRAL_1.getName -> "813361260812",
+    Regions.AP_SOUTHEAST_2.getName -> "544295431143",
+    Regions.GovCloud.getName -> "226302683700"
   )
 }